+++ package.json
@@ -14,6 +14,7 @@
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-toast": "^1.1.5",
+    "@reduxjs/toolkit": "^2.2.5",
     "antd": "^5.18.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
@@ -23,6 +24,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.2.0",
+    "react-redux": "^9.1.2",
     "react-responsive": "^10.0.0",
     "react-toastify": "^10.0.5",
     "tailwind-merge": "^2.3.0",
