+++ src/app/(components)/Providers.tsx
@@ -1,6 +1,10 @@
+"use client";
 import React from "react";
 import dynamic from "next/dynamic";
 import Footer from "./Footer";
+import { Provider } from "react-redux";
+import store from "../redux/store";
+
 // THIS IS STACK OVERFLOW HACK.
 const ResponsiveHeader = dynamic(() => import("./ResponsiveHeader"), {
   ssr: false,
@@ -12,12 +16,12 @@ const ResponsiveFooter = dynamic(() => import("./ResponsiveFooter"), {
 
 const Providers = ({ children }: { children: React.ReactNode }) => {
   return (
-    <div>
+    <Provider store={store}>
       <ResponsiveHeader />
       {children}
       {/* You can add our footer directly now */}
       <ResponsiveFooter />
-    </div>
+    </Provider>
   );
 };
 export default Providers;
