+++ src/app/(components)/ResponsiveFooter.tsx
@@ -3,11 +3,11 @@ import React from "react";
 import { useMediaQuery } from "react-responsive";
 import Footer from "./Footer";
 import MobileFooter from "./MobileFooter";
+import useMediaCustom from "@/hooks/useMediaCustom";
 
 const ResponsiveFooter = () => {
-  const isDesktopOrLaptop = useMediaQuery({
-    query: "(min-width: 1000px)",
-  });
+  const isDesktopOrLaptop = useMediaCustom();
+
   return isDesktopOrLaptop ? <Footer /> : <MobileFooter />;
 };
