+++ src/app/(components)/ResponsiveHeader.tsx
@@ -3,11 +3,10 @@ import React from "react";
 import { useMediaQuery } from "react-responsive";
 import DesktopHeader from "./DesktopHeader";
 import MobileSidebar from "./MobileSidebar";
+import useMediaCustom from "@/hooks/useMediaCustom";
 
 const ResponsiveHeader = () => {
-  const isDesktopOrLaptop = useMediaQuery({
-    query: "(min-width: 1000px)",
-  });
+  const isDesktopOrLaptop = useMediaCustom();
   return isDesktopOrLaptop ? <DesktopHeader /> : <MobileSidebar />;
 };
