+++ src/app/redux/store.js
@@ -1,0 +1,12 @@
+"use client"
+import wishlistDataSlice from "./wishlistDataSlice";
+
+const { configureStore } = require("@reduxjs/toolkit");
+
+const store = configureStore({
+    reducer:{
+        wishlist:wishlistDataSlice
+    }
+});
+
+export default store;
