+++ src/app/wishlist/page.tsx
@@ -1,9 +1,11 @@
+"use client";
 import { Link } from "lucide-react";
 import React from "react";
 import ProductCard from "../(components)/ProductCard";
 import WishlistCard from "../(components)/WishlistCard";
+import { useSelector } from "react-redux";
 
-const page = () => {
+const WishlistPage = () => {
   const landingPageProductData = [
     {
       imageUrl:
@@ -30,26 +32,42 @@ const page = () => {
       productPrice: "2,399",
     },
   ];
-  const arr: number[] = [1, 2, 3, 4, 5, 6, 7, 8];
+  const wishlistData = useSelector(
+    (store: any) => store.wishlist.wishlistArray
+  );
+  console.log(wishlistData);
+  //const arr: number[] = [1, 2, 3, 4, 5, 6, 7, 8];
   return (
-    <div className="flex flex-col justify-center items-center">
+    <div className="flex flex-col justify-start items-center min-h-52">
       <p className="uppercase tracking-widester mt-6 text-md md:text-lg underline">
         MY wishlist
       </p>
-      <div className="flex flex-col justify-center items-start mt-6">
-        <div className="flex flex-wrap gap-2 justify-center items-start">
-          {landingPageProductData.map((data, index) => (
-            <WishlistCard key={index} data={data} />
-          ))}
+      {wishlistData?.length > 0 ? (
+        <div className="flex flex-col justify-center items-start mt-6">
+          <div className="flex flex-wrap gap-2 justify-center items-start">
+            {wishlistData.map((data: any, index: number) => (
+              <WishlistCard
+                key={index}
+                data={{
+                  imageUrl: data.imageUrl[0],
+                  productPrice: data.price,
+                  productTitle: data.title,
+                }}
+              />
+            ))}
+          </div>
         </div>
-        <div className="flex flex-wrap gap-2 justify-center items-start">
-          {landingPageProductData.map((data, index) => (
-            <WishlistCard key={index} data={data} />
-          ))}
-        </div>
-      </div>
+      ) : (
+        <p className="text-xl font-euclid mt-24 uppercase text-center">
+          Your wishlist is empty
+          <br />
+          <span className="text-sm text-center">
+            (but it doesn’t have to be)
+          </span>
+        </p>
+      )}
     </div>
   );
 };
 
-export default page;
+export default WishlistPage;
