+++ src/hooks/useMediaCustom.tsx
@@ -1,0 +1,11 @@
+import React, { useEffect, useState } from "react";
+import { useMediaQuery } from "react-responsive";
+
+const useMediaCustom = () => {
+  const isDesktopOrLaptop = useMediaQuery({
+    query: "(min-width: 1000px)",
+  });
+  return isDesktopOrLaptop;
+};
+
+export default useMediaCustom;
